convert cupy.ceil array to int for easier mathing

+++ pyqtgraph/graphicsItems/ImageItem.py
@@ -523,7 +523,7 @@ class ImageItem(GraphicsObject):
                 return None, None
             if stepData.dtype.kind in "ui":
                 # For integer data, we select the bins carefully to avoid aliasing
-                step = self._xp.ceil((mx - mn) / 500.)
+                step = int(self._xp.ceil((mx - mn) / 500.))
                 bins = []
                 if step > 0.0:
                     bins = self._xp.arange(mn, mx + 1.01 * step, step, dtype=self._xp.int)
